+++ net/net-w32.c
@@ -73,7 +73,7 @@ int net_init(void) {
 }
 
 
-struct net_tcpsocket *net_connect(unsigned long ipaddr, int port) {
+struct net_tcpsocket *net_connect(unsigned long ipaddr, unsigned short port) {
   struct net_tcpsocket *resultsock;
   resultsock = malloc(sizeof(struct net_tcpsocket));
   if (resultsock == NULL) return(NULL);
@@ -102,7 +102,7 @@ struct net_tcpsocket *net_connect(unsigned long ipaddr, int port) {
 
 /* Sends data on socket 'socket'.
    Returns the number of bytes sent on success, and <0 otherwise. The error code can be translated into a human error message via libtcp_strerr(). */
-int net_send(struct net_tcpsocket *socket, char *line, int len) {
+int net_send(struct net_tcpsocket *socket, char *line, long len) {
   int res;
   int status = 0;
   int *statusptr = &status;
@@ -116,7 +116,7 @@ int net_send(struct net_tcpsocket *socket, char *line, int len) {
 
 /* Reads data from socket 'sock' and write it into buffer 'buff', until end of connection. Will fall into error if the amount of data is bigger than 'maxlen' bytes.
 Returns the amount of data read (in bytes) on success, or a negative value otherwise. The error code can be translated into a human error message via libtcp_strerr(). */
-int net_recv(struct net_tcpsocket *socket, char *buff, int maxlen) {
+int net_recv(struct net_tcpsocket *socket, char *buff, long maxlen) {
   int i;
   /* call this to let WatTCP hanle its internal stuff */
   if (tcp_tick(socket->sock) == 0) return(-1);
